requirements.txt
accelerate==1.2.1
boto3>=1.28.43
blinker==1.9.0
Brotli>=1.1.0
certifi==2024.12.14
charset-normalizer==3.4.0
click>=8.1.7
fast-langdetect>=0.2.3
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2024.12.0
huggingface-hub==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
loguru>=0.6.0
MarkupSafe==3.2.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy>=1.21.6,<2.0.0
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20231228
pillow>=10.0.0
psutil==6.1.1
PyMuPDF>=1.24.9
pydantic>=2.7.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn>=1.0.2
setuptools==75.6.0
sympy==1.13.1
tokenizers==0.21.0
torch>=2.0.0
tqdm==4.67.1
transformers>=4.35.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3

# The requirements.txt must ensure that only necessary external dependencies are introduced. If there are new dependencies to add, please contact the project administrator.
